sales_dashboard/requirements.txt
streamlit
pandas
plotly
python-dotenv
snowflake-connector-python